// frontend/src/api/generated/models/roomDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Backend API
 * A simple ASP.NET Core Web API template with best practices for building scalable and maintainable applications.
 * OpenAPI spec version: v1
 */

export interface RoomDto {
  id: number;
  name: string;
  capacity: number;
  locationId: number;
  locationName: string;
  /** @nullable */
  purpose?: string | null;
  /** @nullable */
  imagePath?: string | null;
  createdAt: string;
}